refactor: renamed config pkg
