+++ worker/request.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
-	"scristobal/botatobot/cfg"
+	"scristobal/botatobot/config"
 	"scristobal/botatobot/tasks"
 
 	"github.com/go-telegram/bot/models"
@@ -71,7 +71,7 @@ func (r *Request) MarshalJSON() ([]byte, error) {
 
 func (r *Request) SaveToDisk() error {
 
-	err := os.MkdirAll(cfg.OUTPUT_PATH, 0755)
+	err := os.MkdirAll(config.OUTPUT_PATH, 0755)
 
 	if err != nil {
 		return fmt.Errorf("failed to create output directory: %s", err)
@@ -83,7 +83,7 @@ func (r *Request) SaveToDisk() error {
 		return fmt.Errorf("failed to get result: %s", err)
 	}
 
-	imgFilePath := filepath.Join(cfg.OUTPUT_PATH, fmt.Sprintf("%s.png", r.Id()))
+	imgFilePath := filepath.Join(config.OUTPUT_PATH, fmt.Sprintf("%s.png", r.Id()))
 
 	err = os.WriteFile(imgFilePath, img, 0644)
 
@@ -97,7 +97,7 @@ func (r *Request) SaveToDisk() error {
 		return fmt.Errorf("failed to serialize job parameters: %s", err)
 	}
 
-	jsonFilePath := filepath.Join(cfg.OUTPUT_PATH, fmt.Sprintf("%s.json", r.Id()))
+	jsonFilePath := filepath.Join(config.OUTPUT_PATH, fmt.Sprintf("%s.json", r.Id()))
 
 	err = os.WriteFile(jsonFilePath, content, 0644)
